Finalized trips Controller

+++ eztravel-backend/eztravel-backend/Features/Trips/Views/TripResponse.cs
@@ -16,7 +16,7 @@ public class TripResponse
     
     public int NumberOfPeople { get; set; }
     
-    public decimal Budget { get; set; }
+    public double Budget { get; set; }
     
     public List<TransportView> Transports { get; set; }
